fix error for opencv 4 users
If you using opencv4 the import Stasm throws an error, that it can't find the opencv 3 bindings, when they not installed. And I didn't see any other use of the function somewhere else, so I think its the easiest to just remove it :). Great tool by the way, thanks for sharing